+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+(unreleased)
++++++++++++++++++++
+* `vm/vmss create`: fix a bug that the storage sku is not applied on data disks coming with an image
+
 2.0.13 (2017-08-28)
 +++++++++++++++++++
 * `vmss get-instance-view`: Fix issue where extra, erroneous information was displayed when using `--instance-id *`